+++ test/src/3d/scene/ui/widget/textarea/TextareaTest.cpp
@@ -12,7 +12,7 @@ void TextareaTest::textCut() {
 
     std::string textValue = "mmmmmm"; //textarea can only display 'mmm' on one line
     uiRenderer->onMouseMove(1.0f, 1.0f); //move mouse over textarea
-    uiRenderer->onKeyPress((int)InputDeviceKey::MOUSE_LEFT); //activate textarea
+    uiRenderer->onKeyPress(InputDeviceKey::MOUSE_LEFT); //activate textarea
     for (char textLetter : textValue) {
         uiRenderer->onChar(static_cast<char32_t>(textLetter));
     }
@@ -22,9 +22,9 @@ void TextareaTest::textCut() {
 
     float endOfLinePosX = textarea->getWidth() - (float)textarea->getOutline().rightWidth - 10.0f /* scrollbar width */ - TextFieldConst::TEXT_AND_SCROLLBAR_SHIFT;
     uiRenderer->onMouseMove(endOfLinePosX, 1.0f); //move mouse at end of first line
-    uiRenderer->onKeyPress((int)InputDeviceKey::MOUSE_LEFT); //place cursor at end of first line
+    uiRenderer->onKeyPress(InputDeviceKey::MOUSE_LEFT); //place cursor at end of first line
     for (std::size_t i = 0; i < 3; ++i) {
-        uiRenderer->onKeyPress((int)InputDeviceKey::DELETE_KEY);
+        uiRenderer->onKeyPress(InputDeviceKey::DELETE_KEY);
     }
     AssertHelper::assertUnsignedIntEquals(textarea->getTextWidget().getCutTextLines().size(), 1);
     AssertHelper::assertTrue(textarea->getTextWidget().getCutTextLines()[0].text == WStringConvertA().from_bytes("mmm"));
